+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - General tidy in preparation for cluster creation.
 - Enable HA masters (3).
+- Bump `glitchcrab/terraform-module-proxmox-instance` to 1.6.0
+- Do not power on nodes after creation.
 
 ## [0.1.0] - 2022-12-19